refactor: Migrate contrib download to contrib module (#1119)
* Migrate contrib download code from cli module to contrib module
   - Unify all of the contrib code in a single contrib module for easier factoring out and bookkeeping
* Use DOI URL for all contrib download docstring examples
* Amends PR #1046
